--- source/Esf.DataAccess.Deployment/DeployDb.tests.ps1
@@ -1,2 +1,0 @@
-﻿Describe "DeployDb" {
-}
