+++ source/Esf.DataAccess.DeploymentTests/Esf.DataAccess.DeploymentTests.csproj
@@ -21,7 +21,7 @@
     <DebugSymbols>true</DebugSymbols>
     <DebugType>full</DebugType>
     <Optimize>false</Optimize>
-    <OutputPath>bin\Debug\</OutputPath>
+    <OutputPath>..\..\bin\Esf.DataAccess.DeploymentTests\</OutputPath>
     <DefineConstants>DEBUG;TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
@@ -44,6 +44,10 @@
     <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Esf.DataAccess, Version=1.0.0.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\bin\Esf.DataAccess\Esf.DataAccess.dll</HintPath>
+    </Reference>
     <Reference Include="MongoDB.Bson, Version=2.2.4.26, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\MongoDB.Bson.2.2.4\lib\net45\MongoDB.Bson.dll</HintPath>
       <Private>True</Private>
@@ -94,38 +98,12 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="DbDeploymentConfig.cs" />
-    <Compile Include="EsStateBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="EsStatesRepositoryTests.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\Esf.DataAccess\Esf.DataAccess.csproj">
-      <Project>{37f79771-977a-4e8e-8f38-1a128254e711}</Project>
-      <Name>Esf.DataAccess</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\Esf.DataAccess.Deployment\dbDeployment.config.json">
-      <Link>DeploymentScripts\dbDeployment.config.json</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="..\Esf.DataAccess.Deployment\DeployDb.ps1">
-      <Link>DeploymentScripts\DeployDb.ps1</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="..\Esf.DataAccess.Deployment\StopDbServer.ps1">
-      <Link>DeploymentScripts\StopDbServer.ps1</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
     <None Include="App.config" />
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="..\Esf.DataAccess.Deployment\StopDbServer.js">
-      <Link>DeploymentScripts\StopDbServer.js</Link>
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
   </ItemGroup>
   <Choose>
     <When Condition="'$(VisualStudioVersion)' == '10.0' And '$(IsCodedUITest)' == 'True'">
