+++ source/Esf.DataAccess.DeploymentTests/packages.config
@@ -5,5 +5,12 @@
   <package id="MongoDB.Driver.Core" version="2.2.4" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net461" />
   <package id="NUnit" version="3.2.1" targetFramework="net461" />
+  <package id="NUnit.ConsoleRunner" version="3.4.1" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitProjectLoader" version="3.4.1" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitV2Driver" version="3.4.1" targetFramework="net461" />
+  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.4.1" targetFramework="net461" />
+  <package id="NUnit.Extension.TeamCityEventListener" version="1.0.1" targetFramework="net461" />
+  <package id="NUnit.Extension.VSProjectLoader" version="3.4.1" targetFramework="net461" />
+  <package id="NUnit.Runners" version="3.4.1" targetFramework="net461" />
   <package id="NUnitTestAdapter" version="2.0.0" targetFramework="net461" />
 </packages>
