+++ source/Esf.DataAccess.IntegrationTests/App.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <connectionStrings>
+    <add name="EsFiddleDb" connectionString="mongodb://localhost:27017/esFiddle" />
+  </connectionStrings>
+</configuration>
