+++ source/Esf.DataAccess.IntegrationTests/DbDeploymentConfig.cs
@@ -1,0 +1,28 @@
+﻿using Newtonsoft.Json;
+using System.IO;
+using System.Reflection;
+
+namespace Esf.DataAccess.Tests
+{
+    public class DbDeploymentConfig
+    {
+        public string mongoDbServerDirectory { get; set; }
+        public string esFiddleDbPath { get; set; }
+
+        public static DbDeploymentConfig Load()
+        {
+            string dbDeploymentConfigPath = GetDeploymentScriptPath("dbDeployment.config.json");
+            var jsonContent = File.ReadAllText(dbDeploymentConfigPath);
+            var result = JsonConvert.DeserializeObject<DbDeploymentConfig>(jsonContent);
+            return result;
+
+        }
+
+        public static string GetDeploymentScriptPath(string fileName)
+        {
+            string projectDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+            string result = Path.Combine(projectDirectory, "DeploymentScripts", fileName);
+            return result;
+        }
+    }
+}
