+++ source/Esf.DataAccess.IntegrationTests/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MongoDB.Bson" version="2.2.4" targetFramework="net461" />
+  <package id="MongoDB.Driver" version="2.2.4" targetFramework="net461" />
+  <package id="MongoDB.Driver.Core" version="2.2.4" targetFramework="net461" />
+  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net461" />
+  <package id="NUnit" version="3.4.1" targetFramework="net461" />
+  <package id="NUnitTestAdapter" version="2.0.0" targetFramework="net461" />
+</packages>
